+++ lib/Service/AccountService.php
@@ -39,23 +39,42 @@ class AccountService {
     }
   }
 
-  public function create($name, $type, $currency, $description, $userId) {
+  public function create(
+    $name,
+    $type,
+    $currency,
+    $description,
+    $extraData,
+    $userId
+  ) {
     $account = new Account();
     $account->setName($name);
     $account->setType($type);
     $account->setCurrency($currency);
     $account->setDescription($description);
+    $account->setExtraData($extraData);
+
     $account->setUserId($userId);
     return $this->mapper->insert($account);
   }
 
-  public function update($id, $name, $type, $currency, $description, $userId) {
+  public function update(
+    $id,
+    $name,
+    $type,
+    $currency,
+    $description,
+    $extraData,
+    $userId
+  ) {
     try {
       $account = $this->mapper->find($id, $userId);
       $account->setName($name);
       $account->setType($type);
       $account->setCurrency($currency);
       $account->setDescription($description);
+      $account->setExtraData($extraData);
+
       return $this->mapper->update($account);
     } catch(Exception $e) {
       $this->handleException($e);